+++ src/types/ITokenProps.ts
@@ -1,0 +1,45 @@
+/*-
+ *
+ * Hedera Local Node
+ *
+ * Copyright (C) 2024 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import { CustomFee, Long } from "@hashgraph/sdk";
+
+export interface ITokenProps {
+  tokenName: string;
+  tokenSymbol: string;
+  tokenType: string;
+  supplyType: string;
+  decimals?: number | Long;
+  initialSupply?: number | Long;
+  maxSupply?: number | Long;
+  treasuryKey?: string;
+  adminKey?: string;
+  kycKey?: string;
+  freezeKey?: string;
+  pauseKey?: string;
+  wipeKey?: string;
+  supplyKey?: string;
+  feeScheduleKey?: string;
+  freezeDefault?: boolean;
+  autoRenewAccountId?: string;
+  expirationTime?: string;
+  autoRenewPeriod?: number | Long;
+  tokenMemo?: string;
+  customFees?: CustomFee[];
+}
